requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
discord==2.3.2
discord.py==2.6.4
easy-pil==0.4.0
frozenlist==1.8.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.10
instaloader==4.14.2
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
multidict==6.7.0
pillow==10.4.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
Telethon==1.41.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
